VBFinal/backend/requirements.txt
asgiref==3.11.0
channels==4.1.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.5
daphne==4.1.2
defusedxml==0.7.1
Django==6.0.1
dj-database-url==2.1.0
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-yasg==1.21.12
gunicorn==25.1.0
idna==3.11
inflection==0.5.1
Jinja2==3.1.6
MarkupSafe==2.1.5
oauthlib==3.3.1
packaging==25.0
pillow==12.0.0
psycopg2-binary==2.9.11
pycparser==3.0
PyJWT==2.11.0
python-dotenv==1.2.1
python3-openid==3.2.0
pytz==2025.2
PyYAML==6.0.3
requests==2.32.5
requests-oauthlib==2.0.0
setuptools==70.2.0
social-auth-app-django==5.7.0
social-auth-core==4.8.5
sqlparse==0.5.5
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.6.3
whitenoise==6.12.0
